TarIndex 2GB problem fixed
git-svn-id: https://mobac.svn.sourceforge.net/svnroot/mobac/trunk@665 83dc79ef-cac3-4f51-90f7-e9e68c987b9c

+++ TAC/src/tac/tar/TarArchive.java
@@ -35,7 +35,7 @@ public class TarArchive {
 		this.baseDir = baseDir;
 	}
 
-	public int getTarFilePos() {
+	public long getTarFilePos() {
 		return tarFileStream.getBytesWritten();
 	}
